requirements.txt
yfinance==0.2.65
pandas==2.3.1
numpy==2.3.2
gspread==6.2.1
google-auth-oauthlib==1.2.2
python-dateutil==2.9.0.post0
RapidFuzz==3.13.0
openpyxl==3.1.5
requests==2.32.4